+++ packages/wonder-blocks-button/src/components/button-core.tsx
@@ -110,6 +110,12 @@ function ButtonCore(props: Props) {
                         (pressed
                             ? buttonStyles.active
                             : (hovered || focused) && buttonStyles.focus),
+                    kind === "tertiary" &&
+                        disabled &&
+                        focused && [
+                            buttonStyles.focus,
+                            buttonStyles.disabledFocus,
+                        ],
                 ]}
             >
                 {icon && (
@@ -410,25 +416,12 @@ const _generateStyles = (
             disabled: {
                 color: light ? fadedColor : theme.color.text.disabled,
                 cursor: "default",
-                ":focus": {
-                    ":after": {
-                        content: "''",
-                        position: "absolute",
-                        height: 2,
-                        width: `calc(100% - ${iconWidth}px)`,
-                        right: 0,
-                        /**
-                         * For tertiary buttons, underline is applied to the
-                         * internal `span` element, So we need to calculate the
-                         * offset to center the line in the button (right after
-                         * the span).
-                         */
-                        bottom: `calc(50% - 11px)`,
-                        background: light
-                            ? theme.color.bg.primary
-                            : theme.color.bg.primaryDisabled,
-                        borderRadius: theme.border.radius.xSmall,
-                    },
+            },
+            disabledFocus: {
+                ":after": {
+                    background: light
+                        ? theme.color.bg.primary
+                        : theme.color.bg.primaryDisabled,
                 },
             },
         };