+++ packages/wonder-blocks-dropdown/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @khanacademy/wonder-blocks-dropdown
 
+## 3.0.16
+
+### Patch Changes
+
+-   Updated dependencies [d4c412b5]
+    -   @khanacademy/wonder-blocks-core@5.2.3
+    -   @khanacademy/wonder-blocks-clickable@3.0.13
+    -   @khanacademy/wonder-blocks-icon@2.0.13
+    -   @khanacademy/wonder-blocks-layout@2.0.13
+    -   @khanacademy/wonder-blocks-modal@4.0.14
+    -   @khanacademy/wonder-blocks-search-field@2.1.2
+    -   @khanacademy/wonder-blocks-typography@2.0.13
+
+## 3.0.15
+
+### Patch Changes
+
+-   Updated dependencies [64a188e3]
+    -   @khanacademy/wonder-blocks-core@5.2.2
+    -   @khanacademy/wonder-blocks-clickable@3.0.12
+    -   @khanacademy/wonder-blocks-icon@2.0.12
+    -   @khanacademy/wonder-blocks-layout@2.0.12
+    -   @khanacademy/wonder-blocks-modal@4.0.13
+    -   @khanacademy/wonder-blocks-search-field@2.1.1
+    -   @khanacademy/wonder-blocks-typography@2.0.12
+
+## 3.0.14
+
+### Patch Changes
+
+-   Updated dependencies [3f86013b]
+    -   @khanacademy/wonder-blocks-search-field@2.1.0
+
 ## 3.0.13
 
 ### Patch Changes