+++ packages/wonder-blocks-form/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @khanacademy/wonder-blocks-form
 
+## 4.2.2
+
+### Patch Changes
+
+-   Updated dependencies [d4c412b5]
+    -   @khanacademy/wonder-blocks-core@5.2.3
+    -   @khanacademy/wonder-blocks-clickable@3.0.13
+    -   @khanacademy/wonder-blocks-icon@2.0.13
+    -   @khanacademy/wonder-blocks-layout@2.0.13
+    -   @khanacademy/wonder-blocks-typography@2.0.13
+
+## 4.2.1
+
+### Patch Changes
+
+-   Updated dependencies [64a188e3]
+    -   @khanacademy/wonder-blocks-core@5.2.2
+    -   @khanacademy/wonder-blocks-clickable@3.0.12
+    -   @khanacademy/wonder-blocks-icon@2.0.12
+    -   @khanacademy/wonder-blocks-layout@2.0.12
+    -   @khanacademy/wonder-blocks-typography@2.0.12
+
+## 4.2.0
+
+### Minor Changes
+
+-   3f86013b: Add autocomplete prop to SearchField and TextField
+
 ## 4.1.3
 
 ### Patch Changes