+++ packages/wonder-blocks-popover/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @khanacademy/wonder-blocks-popover
 
+## 2.1.2
+
+### Patch Changes
+
+-   Updated dependencies [d4c412b5]
+    -   @khanacademy/wonder-blocks-core@5.2.3
+    -   @khanacademy/wonder-blocks-icon@2.0.13
+    -   @khanacademy/wonder-blocks-icon-button@4.0.13
+    -   @khanacademy/wonder-blocks-modal@4.0.14
+    -   @khanacademy/wonder-blocks-tooltip@2.0.14
+    -   @khanacademy/wonder-blocks-typography@2.0.13
+
+## 2.1.1
+
+### Patch Changes
+
+-   Updated dependencies [64a188e3]
+    -   @khanacademy/wonder-blocks-core@5.2.2
+    -   @khanacademy/wonder-blocks-icon@2.0.12
+    -   @khanacademy/wonder-blocks-icon-button@4.0.12
+    -   @khanacademy/wonder-blocks-modal@4.0.13
+    -   @khanacademy/wonder-blocks-tooltip@2.0.13
+    -   @khanacademy/wonder-blocks-typography@2.0.12
+
 ## 2.1.0
 
 ### Minor Changes