+++ packages/wonder-blocks-progress-spinner/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-progress-spinner
 
+## 2.0.13
+
+### Patch Changes
+
+-   Updated dependencies [d4c412b5]
+    -   @khanacademy/wonder-blocks-core@5.2.3
+
+## 2.0.12
+
+### Patch Changes
+
+-   Updated dependencies [64a188e3]
+    -   @khanacademy/wonder-blocks-core@5.2.2
+
 ## 2.0.11
 
 ### Patch Changes